Added extensive and interactive documentation comments to all publicly visible declarations, as well as certain private constants and members for context.
Several previous feature changes were determined to violate the ethos for the development of this language, in particular, that '...statements should only utilize up to one expression...'. Such changes included being able to pass an array to the System Args statement, and giving the Goto and Gosub statements a second 'fallback' label. Such changes have been reverted.

Also, a SystemYield statement was introduced to take over the functionaliy of the SystemCall statement, which now no longer adds values to the system output queue.

Finally, the 'L' dynamic literal was reintroduced to give developers quick access to the length of the program arguments string.